refactor: remove `vite-tsconfig-paths`
