+++ .changeset/green-flies-tan.md
@@ -1,0 +1,7 @@
+---
+"astro-reading-time": patch
+"astro-mermaid": patch
+"@mogeko/typography": patch
+---
+
+Remove `vite-tsconfig-paths`.
