+++ packages/astro-mermaid/package.json
@@ -40,7 +40,6 @@
     "tsup": "^8.1.0",
     "typescript": "^5.5.2",
     "unified": "^11.0.5",
-    "vite-tsconfig-paths": "^4.3.2",
     "vitest": "^1.6.0"
   }
 }
