+++ packages/astro-reading-time/package.json
@@ -50,7 +50,6 @@
     "tsconfig": "workspace:^",
     "tsup": "^8.1.0",
     "typescript": "^5.5.2",
-    "vite-tsconfig-paths": "^4.3.2",
     "vitest": "^1.6.0"
   }
 }
