+++ packages/astro-reading-time/vitest.config.ts
@@ -1,4 +1,3 @@
-import tsconfigPaths from "vite-tsconfig-paths";
 import { defineConfig } from "vitest/config";
 
 export default defineConfig({
@@ -10,5 +9,7 @@ export default defineConfig({
       provider: "v8",
     },
   },
-  plugins: [tsconfigPaths()],
+  resolve: {
+    alias: { "@": new URL(".", import.meta.url).pathname },
+  },
 });
